config.sh agora verifica se eh root

+++ config.sh
@@ -1,11 +1,18 @@
 #!/bin/sh
 
+usr=$(whoami)
+if [[ "${usr}" != "root" ]]; then
+	echo "Precisa ser root para executar esse script"
+	echo "Rode 'sudo -i' e tente novamente"
+	exit 1
+fi
+
 echo "enable_uart=1" >> /boot/config.txt
 echo "dtoverlay=pi3-miniuart-bt" >> /boot/config.txt
-chmod 666 /dev/ttyAMA0
 systemctl stop serial-getty@ttyS0.service
 systemctl disable serial-getty@ttyS0.service
 sed -i 's/console=serial0,115200 //g' /boot/cmdline.txt
+pip install pymcuprog
 
 echo "i2c-dev" >> /etc/modules
 echo "i2c-bcm2708" >> /etc/modules
